third_party/WebKit/LayoutTests/vibration/vibration-exceptions.html

<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="vibration-utils.js"></script>
</head>
<body>
<script>
description('Tests that bad input throws exceptions in the Vibration API.');

shouldThrow("navigator.vibrate();");
shouldThrow("navigator.vibrate(1, 2);");
shouldThrow("navigator.vibrate([1], [2]);");

</script>
</body>
</html>
